+++ internal/models/contact_point.go
@@ -10,7 +10,7 @@ import (
 type ContactPoint struct {
 	ID            [16]byte               `json:"id"`
 	Name          string                 `json:"name"`
-	UserID        int64                  `json:"user_id"`
+	UserID        int                    `json:"user_id"`
 	Type          string                 `json:"type"`
 	Configuration map[string]interface{} `json:"configuration"` // Stored as string in DB
 	Status        string                 `json:"status"`
@@ -20,7 +20,7 @@ type ContactPoint struct {
 
 type ContactPointCreate struct {
 	Name          string                 `json:"name" binding:"required"`
-	UserID        int64                  `json:"user_id" binding:"required"`
+	UserID        int                    `json:"user_id" binding:"required"`
 	Type          string                 `json:"type" binding:"required"`
 	Configuration map[string]interface{} `json:"configuration" binding:"required"`
 }
@@ -28,7 +28,7 @@ type ContactPointCreate struct {
 type ContactPointUpdate struct {
 	ID            string                 `json:"id" binding:"required"`
 	Name          string                 `json:"name,omitempty"`
-	UserID        *int64                 `json:"user_id,omitempty"`
+	UserID        *int                   `json:"user_id,omitempty"`
 	Type          string                 `json:"type,omitempty"`
 	Configuration map[string]interface{} `json:"configuration,omitempty"`
 	Status        string                 `json:"status,omitempty"`
